Big changes in architecture
Move to object-oriented
Add an large AbstructLayer of underlying interface

+++ src/Kernel/Global.h
@@ -10,7 +10,7 @@
 class IntVec4
 {
 public:
-	long long x,y,z;
+	int x,y,z;
     int dim;
 };
